feat(hyd): refactored flaps motors
